fix: fix utils export [MA-1404]
