+++ packages/analytics/metric-cards/src/types/index.ts
@@ -1,7 +1,1 @@
-// Export all types and interfaces from this index.ts
-// The actual types and interfaces should be contained in separate files within this folder.
-
-// Example:
-// export * from './component-types'
-
 export * from './metric-card'
